% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mccall_lab.R
\name{scale_mccall_lab}
\alias{scale_mccall_lab}
\alias{mccall_lab_pal}
\alias{scale_color_mccall_lab}
\alias{scale_colour_mccall_lab}
\alias{scale_fill_mccall_lab}
\title{Color palette for McCall Lab plots
8 color scale created by [Martin Krzywinski's color blind palettes](http://mkweb.bcgsc.ca/colorblind/palettes.mhtml#page-container), adapted from [Wong, B. (2011) Nature Methods 8:441](https://www.nature.com/articles/nmeth.1618).}
\usage{
mccall_lab_pal()

scale_color_mccall_lab(...)

scale_colour_mccall_lab(...)

scale_fill_mccall_lab(...)
}
\arguments{
\item{...}{Further arguments passed to other methods}
}
\description{
Color palette for McCall Lab plots
8 color scale created by [Martin Krzywinski's color blind palettes](http://mkweb.bcgsc.ca/colorblind/palettes.mhtml#page-container), adapted from [Wong, B. (2011) Nature Methods 8:441](https://www.nature.com/articles/nmeth.1618).
}
\references{
Krzywinski, M. Designing for Color Blindess - Martin Krzywinski / Genome Sciences Center. Retrieved March 5, 2021, from http://mkweb.bcgsc.ca/colorblind/index.mhtml#page-container
Wong, B. (2011). Color blindness. In Nature Methods (Vol. 8, Issue 6, p. 441). https://doi.org/10.1038/nmeth.1618
}
\seealso{
[ggplot2::scale_color_discrete][ggplot2::scale_fill_discrete]
}
