% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mccall_lab.R
\name{theme_mccall_lab}
\alias{theme_mccall_lab}
\title{McCall Lab Theme}
\usage{
theme_mccall_lab(
  text_size = 17,
  title_size = 21,
  text_color = "black",
  panel_grid_linetype = "18",
  base_theme = ggplot2::theme_linedraw()
)
}
\arguments{
\item{text_size}{Font size for axis text. Legend text will scale by 0.8. Defaults to 17pt.}

\item{title_size}{Font size for title text. Subtitle will scale by 0.9 and axis title will scale by 0.8. Defaults to 21pt.}

\item{text_color}{Color for all text in plot. Defaults to black.}

\item{panel_grid_linetype}{Line type for grid lines in plot. Defaults to 18, a dotted line with 1px on followed by 8px off.}

\item{base_theme}{Underlying theme to build from. Defaults to linedraw, displaying a margin outline for the plot and gridlines.}
}
\description{
A theme for [ggplot2] with color blind friendly palettes for use
members of the McCall Lab.
}
\examples{
library(ggplot2)
library(palmerpenguins)


ggplot(penguins, aes(x = bill_length_mm, y = bill_depth_mm, color = species)) +
    scale_color_mccall_lab() +
    theme_mccall_lab()

}
\seealso{
[ggplot2::theme]
}
