% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_mccall_lab.R
\name{scale_mccall_lab}
\alias{scale_mccall_lab}
\alias{mccall_lab_pal}
\alias{scale_color_mccall_lab}
\alias{scale_colour_mccall_lab}
\alias{scale_fill_mccall_lab}
\title{Color palette for McCall Lab plots
8 color scale created by [Martin Krzywinski's color blind palettes](http://mkweb.bcgsc.ca/colorblind/palettes.mhtml#page-container), adapted from [Wong, B. (2011) Nature Methods 8:441](https://www.nature.com/articles/nmeth.1618).}
\usage{
mccall_lab_pal(palette = c("martink8", "martink12", "martink15", "martink48"))

scale_color_mccall_lab(
  palette = c("martink8", "martink12", "martink15", "martink48"),
  ...
)

scale_colour_mccall_lab(
  palette = c("martink8", "martink12", "martink15", "martink48"),
  ...
)

scale_fill_mccall_lab(
  palette = c("martink8", "martink12", "martink15", "martink48"),
  ...
)
}
\arguments{
\item{palette}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take (e.g., \code{\link[scales:hue_pal]{scales::hue_pal()}}).}

\item{...}{Further arguments passed to other methods}

\item{palette=}{Palette for color scale. Takes on the values martink8,
martink12, martink15, or martink48. See reference for images.}
}
\description{
Color palette for McCall Lab plots
8 color scale created by [Martin Krzywinski's color blind palettes](http://mkweb.bcgsc.ca/colorblind/palettes.mhtml#page-container), adapted from [Wong, B. (2011) Nature Methods 8:441](https://www.nature.com/articles/nmeth.1618).
}
\references{
Krzywinski, M. Designing for Color Blindess - Martin Krzywinski / Genome Sciences Center. http://mkweb.bcgsc.ca/colorblind/index.mhtml#page-container
Wong, B. (2011). Color blindness. In Nature Methods (Vol. 8, Issue 6, p. 441). https://doi.org/10.1038/nmeth.1618
}
\seealso{
[ggplot2::scale_color_discrete][ggplot2::scale_fill_discrete]
}
